+++ pgcli/packages/sqlcompletion.py
@@ -444,7 +444,7 @@ def suggest_based_on_last_token(token, stmt):
             return (Role(),)
         return (Keyword(),)
 
-    elif token_v in ("select", "where", "having", "order by", "distinct"):
+    elif token_v in ("select", "where", "having", "group by", "order by", "distinct"):
         return _suggest_expression(token_v, stmt)
     elif token_v == "as":
         # Don't suggest anything for aliases